==> 01_Initial_Data_Setup.ipynb <==
# 1. Setup All Workshop Dependencies

## _Note:  This Notebook Will Take A Few Minutes To Complete._

## _Please Be Patient._

!python --version

!pip list

## Package Installs

!pip install --disable-pip-version-check -q pip --upgrade > /dev/null
!pip install --disable-pip-version-check -q wrapt --upgrade > /dev/null
!pip install --disable-pip-version-check -q awscli boto3

!pip install --disable-pip-version-check -q sagemaker
!pip install --disable-pip-version-check -q smdebug
!pip install --disable-pip-version-check -q sagemaker-experiments

!pip install --disable-pip-version-check -q PyAthena

!pip install --disable-pip-version-check -q awswrangler

!conda install -y zip

!pip install --disable-pip-version-check -q matplotlib

!pip install --disable-pip-version-check -q seaborn

packages_installed = True

%store packages_installed

# 2. Create S3 Bucket

import boto3
import sagemaker

session = boto3.session.Session()
region = session.region_name
sagemaker_session = sagemaker.Session()
bucket = sagemaker_session.default_bucket()
role = sagemaker.get_execution_role()
account_id = boto3.client("sts").get_caller_identity().get("Account")

s3 = boto3.Session().client(service_name="s3", region_name=region)

setup_s3_bucket_passed = False

print(f"Default bucket: {bucket}")

from botocore.client import ClientError

response = None

try:
    response = s3.head_bucket(Bucket=bucket)
    print(response)
    setup_s3_bucket_passed = True
except ClientError as e:
    print(f"[ERROR] Cannot find bucket {bucket} due to {e}.")

# Define S3 paths
s3_csv_private_path = f"s3://{bucket}/airline-delay-cause/csv"

# Store variables using %store
%store setup_s3_bucket_passed
%store region
%store role
%store account_id


%store

# 3. Copy CSV to S3 Bucket

%store -r packages_installed

try:
    packages_installed
except NameError:
    print("+++++++++++++++++++++++++++++++")
    print("[ERROR] Package Installation Incomplete.")
    print("+++++++++++++++++++++++++++++++")

%store -r setup_s3_bucket_passed

try:
    setup_s3_bucket_passed
except NameError:
    print("+++++++++++++++++++++++++++++++")
    print("[ERROR] S3 bucket not initilized")
    print("+++++++++++++++++++++++++++++++")

s3_csv_private_path = "s3://{}/airline-delay-cause/csv".format(bucket)
print(s3_csv_private_path)

!aws s3 cp "data/Airline_Delay_Cause.csv.gz" $s3_csv_private_path/
!aws s3 ls $s3_csv_private_path/

%store s3_csv_private_path

# 4. Create Athena Database Schema

create_base_csv_athena_db = False

%store -r s3_csv_private_path

try:
    s3_csv_private_path
except NameError:
    print("*****************************************************************************")
    print("[ERROR] CSV not copied into S3 Bucket")
    print("*****************************************************************************")

from pyathena import connect
import pandas as pd

database_name = "db_airline_delay_cause"

# Set S3 staging directory -- this is a temporary directory used for Athena queries
s3_staging_dir = "s3://{0}/athena/staging".format(bucket)

# Connect to Athena
conn = connect(region_name=region, s3_staging_dir=s3_staging_dir)

# Create a cursor object
cursor = conn.cursor()

# Execute the query
statement = f"CREATE DATABASE IF NOT EXISTS {database_name}"
cursor.execute(statement)

print(f"Database {database_name} created successfully (if it did not already exist).")

%store s3_staging_dir


## Verify Successful Database Creation

# Execute the query
statement = "SHOW DATABASES"
cursor.execute(statement)

# Fetch results and convert to DataFrame
df_show = pd.DataFrame(cursor.fetchall(), columns=["database_name"])

# Display first 5 rows
print(df_show.head(5))

# Check if database exists
create_base_csv_athena_db = database_name in df_show["database_name"].values

%store create_base_csv_athena_db


# 5. Register S3 CSV with Athena

create_base_csv_athena_table = False

%store -r create_base_csv_athena_db

try:
    create_base_csv_athena_db
except NameError:
    print("++++++++++++++++++++++++++++++++++++++++++++++")
    print("[ERROR] Athena DB was not created")
    print("++++++++++++++++++++++++++++++++++++++++++++++")

%store -r s3_csv_private_path
print(s3_csv_private_path)

#### Dataset columns
- `year`: YYYY format
- `month`: MM format (1-12)
- `carrier`: Code assigned by assigned by US DOT to identify a unique airline carrier.  
- `carrier_name`: Unique airline (carrier) is defined as one holding and reporting under the same DOT certificate regardless of its Code, Name, or holding company/corporation.
- `airport`: A three character alpha-numeric code issued by the U.S. Department of Transportation which is the official designation of the airport.
- `airport_name`: a place from which aircraft operate that usually has paved runways and maintenance facilities and often serves as a terminal
- `arr_flights`: Arrival Flights
- `arr_del15`: Arrival Delay Indicator, 15 Minutes or More Arrival delay equals the difference of the actual arrival time minus the scheduled arrival time. A flight is considered on-time when it arrives less than 15 minutes after its published arrival time.
- `carrier_ct`:	Carrier Count for airline cause of delay
- `weather_ct`:	Weather Count for airline cause of delay
- `nas_ct`:	NAS (National Air System) Count for airline cause of delay
- `security_ct`: Security County for airline cause of delay
- `late_aircraft_ct`: Late Aircraft Delay Count for airline cause of delay
- `arr_cancelled`: flight cancelled
- `arr_diverted`: flight diverted
- `arr_delay`: Difference in minutes between scheduled and actual arrival time. Early arrivals show negative numbers.
- `carrier_delay`: Carrier Delay, in Minutes
- `weather_delay`: Weather Delay, in Minutes
- `nas_delay`: National Air System Delay, in Minutes
- `security_delay`: Security Delay, in Minutes
- `late_aircraft_delay`: Late Aircraft Delay, in Minutes



from pyathena import connect
import pandas as pd

# Set S3 staging directory -- this is a temporary directory used for Athena queries
s3_staging_dir = "s3://{0}/athena/staging".format(bucket)

# Set Athena parameters
database_name = "db_airline_delay_cause"
raw_table_name = "airline_delay_cause_csv_raw"

conn = connect(region_name=region, s3_staging_dir=s3_staging_dir)

# SQL statement to execute
statement = """CREATE EXTERNAL TABLE IF NOT EXISTS {}.{}(
         year int,
         month int,
         carrier string,
         airport string,
         arr_flights int,
         arr_del15 int,
         carrier_ct float,
         weather_ct float,
         nas_ct float,
         security_ct float,
         late_aircraft_ct float,
         arr_cancelled int,
         arr_diverted int,
         arr_delay int,
         carrier_delay int,
         weather_delay int,
         nas_delay int,
         security_delay int,
         late_aircraft_delay int
) ROW FORMAT DELIMITED FIELDS TERMINATED BY ',' LINES TERMINATED BY '\\n' LOCATION '{}'
TBLPROPERTIES ('compressionType'='gzip', 'skip.header.line.count'='1')""".format(
    database_name, raw_table_name, s3_csv_private_path
)

print(statement)

pd.read_sql(statement, conn)

## Verify Table Creation

statement = "SHOW TABLES in {}".format(database_name)

df_show = pd.read_sql(statement, conn)
print(df_show.head(5))

if raw_table_name in df_show.values:
    create_base_csv_athena_table = True

%store create_base_csv_athena_table

## Run a Sample Query

carrier = "9E"

statement = """SELECT * FROM {}.{}
    WHERE carrier = '{}' LIMIT 100""".format(
    database_name, raw_table_name, carrier
)

print(statement)

df = pd.read_sql(statement, conn)
df.head(5)

%store database_name raw_table_name

print()

%store

%%html

<p><b>Shutting down your kernel for this notebook to release resources.</b></p>
<button class="sm-command-button" data-commandlinker-command="kernelmenu:shutdown" style="display:none;">Shutdown Kernel</button>
        
<script>
try {
    els = document.getElementsByClassName("sm-command-button");
    els[0].click();
}
catch(err) {
    // NoOp
}    
</script>

%%javascript

try {
    Jupyter.notebook.save_checkpoint();
    Jupyter.notebook.session.delete();
}
catch(err) {
    // NoOp
}
